#5 ロジックの整理
