+++ app/Services/Google/GoogleDriveService.php
@@ -21,18 +21,17 @@ class GoogleDriveService
     {
         try {
             $client = new Google_Client();
-            
+
             $authType = config('services.google.auth_type', 'service-account');
-            
+
             if ($authType === 'service-account') {
                 $this->initializeServiceAccount($client);
             } else {
                 $this->initializeOAuth($client);
             }
-            
+
             $this->service = new Google_Service_Drive($client);
             $this->folderId = config('services.google.folder_id');
-            
         } catch (\Exception $e) {
             Log::error('Google Drive初期化エラー', [
                 'error' => $e->getMessage(),
@@ -45,7 +44,7 @@ class GoogleDriveService
     private function initializeServiceAccount(Google_Client $client): void
     {
         $credentialsPath = config('services.google.credentials_path');
-        
+
         if (!file_exists($credentialsPath)) {
             throw new GoogleDriveException('認証ファイルが見つかりません: ' . $credentialsPath);
         }
@@ -62,8 +61,8 @@ class GoogleDriveService
     public function uploadFile($file, ?string $subFolder = null): array
     {
         try {
-            $parentId = $subFolder ? 
-                $this->getOrCreateFolder($subFolder) : 
+            $parentId = $subFolder ?
+                $this->getOrCreateFolder($subFolder) :
                 $this->folderId;
 
             $fileMetadata = new \Google_Service_Drive_DriveFile([
@@ -85,7 +84,6 @@ class GoogleDriveService
                 'view_link' => $result->webViewLink,
                 'name' => $result->name
             ];
-
         } catch (\Exception $e) {
             Log::error('ファイルアップロードエラー', [
                 'error' => $e->getMessage(),
@@ -96,29 +94,29 @@ class GoogleDriveService
     }
 
     private function getOrCreateFolder(string $folderName): string
-{
-    $query = sprintf("name = '%s' and mimeType = 'application/vnd.google-apps.folder' and trashed = false", $folderName);
+    {
+        $query = sprintf("name = '%s' and mimeType = 'application/vnd.google-apps.folder' and trashed = false", $folderName);
 
-    $response = $this->service->files->listFiles([
-        'q' => $query,
-        'spaces' => 'drive',
-        'fields' => 'files(id, name)',
-    ]);
+        $response = $this->service->files->listFiles([
+            'q' => $query,
+            'spaces' => 'drive',
+            'fields' => 'files(id, name)',
+        ]);
 
-    if (count($response->files) > 0) {
-        return $response->files[0]->id;
+        if (count($response->files) > 0) {
+            return $response->files[0]->id;
+        }
+
+        $fileMetadata = new \Google_Service_Drive_DriveFile([
+            'name' => $folderName,
+            'mimeType' => 'application/vnd.google-apps.folder',
+            'parents' => [$this->folderId]
+        ]);
+
+        $folder = $this->service->files->create($fileMetadata, [
+            'fields' => 'id'
+        ]);
+
+        return $folder->id;
     }
-
-    $fileMetadata = new \Google_Service_Drive_DriveFile([
-        'name' => $folderName,
-        'mimeType' => 'application/vnd.google-apps.folder',
-        'parents' => [$this->folderId]
-    ]);
-
-    $folder = $this->service->files->create($fileMetadata, [
-        'fields' => 'id'
-    ]);
-
-    return $folder->id;
-}
 }
