+++ routes/web.php
@@ -8,15 +8,7 @@ Route::get('/', function () {
 });
 
 
-
 Route::get('/upload', function () {
     return view('files.upload');
 })->name('files.upload');
-
-Route::post('/upload', [FileController::class, 'upload'])->name('files.store');
-
-
-Route::get('/simple-upload', function () {
-    return view('files.simpleUpload');
-})->name('files.simple-upload');
-Route::post('/simple-upload', [FileController::class, 'simpleUpload'])->name('files.simple-store');
+Route::post('/upload', [FileController::class, 'fileUpload'])->name('files.store');
